synchronize offset get and data write and refactor package structure
