+++ server/src/main/java/org/apache/iotdb/db/metadata/logfile/MetadataOperationType.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.metadata;
+package org.apache.iotdb.db.metadata.logfile;
 
 public class MetadataOperationType {
